feat(task): backlog par projet, scopé dans les queries (M2)
Ajoute le module `task` : migration `tasks` / `task_notes`, queries sqlc et
la feature handler/service/store complète (list, get, create, update,
add_note, archive).

Isolation — la promesse centrale du produit :

- `team_id` ET `project_id` figurent dans chaque query, y compris
  l'insertion d'une note, alimentée par un SELECT scopé sur la tâche ;
- une clé étrangère composite `(project_id, team_id)` vers
  `projects (id, team_id)` rend impossible une tâche dont le `team_id`
  diverge de celui de son projet ;
- toutes les routes exigent un token de portée `project`. Aucune route ne
  prend de projet en paramètre, donc il n'existe aucune surface où un scope
  pourrait être contourné ;
- couvert par des tests d'intégration, dont l'isolation entre deux projets
  d'une même team — le cas qu'un filtrage limité à `team_id` laisserait
  passer.

Autres décisions :

- la réservation du numéro et l'insertion partagent une transaction
  (`Transactor` implémenté dans le store) : une création refusée ne brûle
  pas un numéro de la suite CORE-1, CORE-2, …
- `ClaimNextNumber` renvoyait un `int32` — sqlc n'infère pas `bigint` pour
  l'expression `next_number - 1`. Cast explicite ajouté.
- `ErrNotFound` couvre indifféremment « numéro inexistant », « tâche
  archivée » et « tâche d'un autre projet » : distinguer ces cas dirait à un
  agent quels numéros existent dans un projet auquel il n'a pas accès.
- le listing ne transporte pas la description, qui peut faire des dizaines
  de milliers d'octets par tâche.

+++ sql/queries/projects.sql
@@ -16,9 +16,11 @@ SELECT * FROM projects WHERE team_id = $1 ORDER BY key;
 
 -- ClaimNextNumber réserve le prochain identifiant lisible du projet (FRNT-34).
 -- Le UPDATE ... RETURNING sérialise les appels concurrents sur la ligne du projet.
+-- Le cast en bigint est explicite : sans lui, sqlc infère un int32 pour l'expression
+-- `next_number - 1`, alors que la colonne est un bigint.
 -- name: ClaimNextNumber :one
 UPDATE projects
 SET next_number = next_number + 1,
     updated_at  = now()
 WHERE id = $1 AND team_id = $2
-RETURNING next_number - 1 AS claimed_number;
+RETURNING (next_number - 1)::bigint AS claimed_number;
